Webscraper(mostly) Finished!
I completed the data formatting section for game data, then created the dataframe section where we store all the data.

This was difficult, as it needed to be manipulated a bit to get it how I wanted.

However, at the end of the program we now have a suitable pandas dataframe! All that's left to do is write it to a csv, and analyze in a jupyter notebook.